pyproject.toml
[project]
name = "dl_workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "gitpython>=3.1.44",
    "numpy<2.0.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
]
